wip moving classes to protocol
