+++ distributed/src/main/scala/asobu/distributed/service/Syntax.scala
@@ -4,7 +4,7 @@ import akka.actor.{ActorRef, ActorSystem}
 import akka.util.Timeout
 import asobu.distributed.service.extractors.DRequestExtractor
 import asobu.distributed.{DRequest, DResult}
-import asobu.distributed.{RequestEnricherDefinition, EndpointDefinition, Headers}
+import asobu.distributed.{RequestEnricherDefinition, Headers}
 import asobu.dsl.{ExtractResult, Extractor, ExtractorFunctions}
 import play.api.libs.json.{Json, Reads, Writes}
 import play.api.mvc.{Result, Results}
